+++ Dockerfile
@@ -2,15 +2,15 @@ FROM golang:1.18-bullseye AS builder
 
 COPY . /work
 WORKDIR /work
-RUN useradd pressurecooker
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o app/pressurecooker cmd/main.go
+RUN useradd multicooker
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o app/multicooker cmd/main.go
 
 FROM  alpine
 
-COPY  --from=builder /work/app/pressurecooker /usr/sbin/pressurecooker
+COPY  --from=builder /work/app/multicooker /usr/sbin/multicooker
 COPY --from=builder /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
 COPY --from=builder /etc/passwd /etc/
 RUN chmod 777 /tmp/
-USER pressurecooker
+USER multicooker
 
-ENTRYPOINT ["/usr/sbin/pressurecooker", "-logtostderr"]
+ENTRYPOINT ["/usr/sbin/multicooker", "-logtostderr"]
