+++ chart/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v1
 appVersion: "1.0"
 description: A Helm chart for Kubernetes
-name: pressurecooker
+name: multicooker
 version: 0.1.0
