+++ pkg/multicooker/evicter_evict.go
@@ -1,4 +1,4 @@
-package pressurecooker
+package multicooker
 
 import (
 	"context"
