+++ pkg/multicooker/watcher_type.go
@@ -1,4 +1,4 @@
-package pressurecooker
+package multicooker
 
 import (
 	"time"
